% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chain}
\alias{chain}
\title{Use random walk to get Markov chain  using Rcpp.}
\usage{
chain(sig, n1)
}
\arguments{
\item{sig}{the standard deviation of the  norm distribution (numeric)}

\item{n1}{the number of sample (numeric)}
}
\value{
a Markov chain of size \code{n1} and the accept rate
}
\description{
Use random walk to get Markov chain  using Rcpp.
}
\examples{
\dontrun{
chain(1,1000)
}
}
