% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{eqdist.nn6}
\alias{eqdist.nn6}
\title{A funciton with nn}
\description{
a function used in nn
}
\examples{
\dontrun{
eqdist.nn6 <- function(z,sizes,k){
 boot.obj <- boot(data=z,statistic=Tn62,R=999,
                  sim = "permutation", sizes = sizes,k=k)
 ts = c(boot.obj$t0,boot.obj$t)
 p.value =mean(ts>=ts[1])
 list(statistic=ts[1],p.value=p.value)
}
}
}
