% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mcmc.norm}
\alias{mcmc.norm}
\title{A two dimensional normal Gibbs sampler}
\usage{
mcmc.norm(m, mu1, mu2, sigma1, sigma2, rho)
}
\arguments{
\item{m}{the number of repeat}

\item{mu1}{the mean of the first norm distribution}

\item{mu2}{the mean of the first norm distribution}

\item{sigma1}{the standard deviation of the first norm distribution}

\item{sigma2}{the standard deviation of the first norm distribution}

\item{rho}{the correlation coefficient}
}
\value{
a Gibbs sampler of the distribution
}
\description{
two dimensional normal Gibbs sampler
}
\examples{
\dontrun{
mcmc.norm(1000,1,2,1,4,0.5)
}
}
