% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mixnorm.em}
\alias{mixnorm.em}
\title{using EM algorithm to estimate GMM}
\usage{
mixnorm.em(x1)
}
\arguments{
\item{x1}{the sample to be solved}
}
\value{
a list of \code{mu} and \code{sigma}
}
\description{
using EM algorithm to estimate GMM
}
\examples{
\dontrun{
mixnorm.em(as.matrix(faithful))
}
}
