% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random}
\alias{random}
\title{Generating normal distribution using Rcpp}
\usage{
random(sigma, n)
}
\arguments{
\item{sigma}{the standard deviation of the  norm distribution (numeric)}

\item{n}{the number of sample (numeric)}
}
\value{
a normal distribution sample of size \code{n}
}
\description{
Generating normal distribution using Rcpp
}
\examples{
\dontrun{
random(1,100)
}
}
