\name{StatComp20013-package}
\alias{StatComp20013-package}
\docType{package}
\title{
  This package provides several functions for R and Rcpp.
}
\description{
  four function are provided in this package (\code{mcmc.norm} and \code{mixnorm.em} versus \code{random} and \code{chain})
}
\author{
Zhi ya Chen

Maintainer: zhi ya chen <chenzy16@mail.ustc.edu.cn>
}
\keyword{ package }
