% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Tn62}
\alias{Tn62}
\title{A funciton used for illustration.}
\description{
a function used in nn
}
\examples{
\dontrun{
Tn62 = function(z, ix, sizes,k) {
 n1 = sizes[1] 
 n2 = sizes[2]
 n = n1 + n2
 if(is.vector(z)) {z = data.frame(z,0)}
 z = z[ix, ]
 NN = nn2(data=z, k=k+1) 
 block1 = NN$nn.idx[1:n1,-1]
 block2 = NN$nn.idx[(n1+1):n,-1]
 i1 = sum(block1 < n1 + .5)
 i2 = sum(block2 > n1+.5)
 return((i1 + i2) / (k * n))
}
}
}
